src/importer/pyproject.toml
[tool.poetry]
name = "importer"
version = "0.1.0"
description = ""
authors = ["Zoltan Kohalmy <zoltankohalmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
kafka-python = "^2.0.1"
attrs = "^19.3.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
